+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - [Support to Material styles][issue-142]
+- [Counter component][issue-141]
 
 ## 1.6.0 - 2019-10-14
 
@@ -197,5 +198,5 @@
 [issue-121]:https://github.com/stone-payments/emerald-components-android/issues/121
 [issue-125]:https://github.com/stone-payments/emerald-components-android/issues/125
 [issue-129]:https://github.com/stone-payments/emerald-components-android/issues/129
+[issue-141]:https://github.com/stone-payments/emerald-components-android/issues/141
 [issue-142]:https://github.com/stone-payments/emerald-components-android/issues/142
-